update tests and add backoff
