+++ samples/snippets/requester_pays_test.py
@@ -12,9 +12,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import backoff
 import os
 import tempfile
+import time
 
+from google.api_core.exceptions import GoogleAPIError
 from google.cloud import storage
 import pytest
 
@@ -30,22 +33,19 @@ import storage_get_requester_pays_status
 PROJECT = os.environ["GOOGLE_CLOUD_PROJECT"]
 
 
-def test_enable_requester_pays(requester_pays_bucket, capsys):
-    storage_enable_requester_pays.enable_requester_pays(requester_pays_bucket.name)
-    out, _ = capsys.readouterr()
-    assert f"Requester Pays has been enabled for {requester_pays_bucket.name}" in out
-
-
-def test_disable_requester_pays(requester_pays_bucket, capsys):
-    storage_disable_requester_pays.disable_requester_pays(requester_pays_bucket.name)
-    out, _ = capsys.readouterr()
-    assert f"Requester Pays has been disabled for {requester_pays_bucket.name}" in out
-
-
-def test_get_requester_pays_status(requester_pays_bucket, capsys):
-    storage_get_requester_pays_status.get_requester_pays_status(requester_pays_bucket.name)
-    out, _ = capsys.readouterr()
-    assert f"Requester Pays is disabled for {requester_pays_bucket.name}" in out
+@pytest.fixture(scope="session")
+def requester_pays_bucket():
+    """Yields a bucket used for requester pays tests."""
+    # We use a different bucket from other tests.
+    # The service account for the test needs to have Billing Project Manager role
+    # in order to make changes on buckets with requester pays enabled.
+    rpays_bucket_name = os.environ["REQUESTER_PAYS_TEST_BUCKET"]
+    bucket = storage.Client().bucket(rpays_bucket_name)
+    if not bucket.exists():
+        bucket.create()
+    yield bucket
+    time.sleep(3)
+    bucket.delete(force=True)
 
 
 @pytest.fixture
@@ -57,6 +57,14 @@ def test_blob(requester_pays_bucket):
     return blob
 
 
+@backoff.on_exception(backoff.expo, GoogleAPIError, max_time=60)
+def test_enable_requester_pays(requester_pays_bucket, capsys):
+    storage_enable_requester_pays.enable_requester_pays(requester_pays_bucket.name)
+    out, _ = capsys.readouterr()
+    assert f"Requester Pays has been enabled for {requester_pays_bucket.name}" in out
+
+
+@backoff.on_exception(backoff.expo, GoogleAPIError, max_time=60)
 def test_download_file_requester_pays(requester_pays_bucket, test_blob):
     with tempfile.NamedTemporaryFile() as dest_file:
         storage_download_file_requester_pays.download_file_requester_pays(
@@ -64,3 +72,17 @@ def test_download_file_requester_pays(requester_pays_bucket, test_blob):
         )
 
         assert dest_file.read()
+
+
+@backoff.on_exception(backoff.expo, GoogleAPIError, max_time=60)
+def test_disable_requester_pays(requester_pays_bucket, capsys):
+    storage_disable_requester_pays.disable_requester_pays(requester_pays_bucket.name, PROJECT)
+    out, _ = capsys.readouterr()
+    assert f"Requester Pays has been disabled for {requester_pays_bucket.name}" in out
+
+
+@backoff.on_exception(backoff.expo, GoogleAPIError, max_time=60)
+def test_get_requester_pays_status(requester_pays_bucket, capsys):
+    storage_get_requester_pays_status.get_requester_pays_status(requester_pays_bucket.name)
+    out, _ = capsys.readouterr()
+    assert f"Requester Pays is disabled for {requester_pays_bucket.name}" in out
